refactor(arte-odyssey): demote clamp to internal
clamp は単なる数値ユーティリティで、toPrecision を非公開にした判断と一貫させるため src/internal/ に移動した。

公開 Helpers は React/DOM 拡張用の5つ (cn / mergeRefs / mergeProps / chain / createSafeContext) に統一。
