+++ .changeset/refine-public-helpers.md
@@ -2,7 +2,7 @@
 '@k8o/arte-odyssey': major
 ---
 
-公開している Helpers を整理し、UIライブラリとして必要な6つに絞った。
+公開している Helpers を整理し、UIライブラリとして必要な5つに絞った。
 
 新しい公開 Helpers:
 
@@ -10,12 +10,11 @@
 - `mergeRefs` (新規公開) — 複数のrefを1つの要素に結合する
 - `mergeProps` (新規) — className/style/イベントハンドラを適切にマージしながら複数のpropsを合成する
 - `chain` (新規) — 複数の関数を順番に呼び出す関数を作る
-- `clamp` (`between` をリネーム) — 数値を最小値と最大値の間にクランプする
 - `createSafeContext` (新規) — Provider外でアクセスされた場合に明確にthrowするContextを作成する
 
 廃止または非公開化した Helpers:
 
-- `between` → `clamp` にリネーム
+- `between` → 内部実装 (`src/internal/clamp.ts`) として `clamp` にリネーム
 - `findAllColors` → Code コンポーネント内部に colocate
 - `isInternalRoute` → Anchor 内にインライン化
 - `cast` → NumberField コンポーネント内部に colocate
