+++ apps/docs/src/i18n/types.ts
@@ -252,7 +252,6 @@ export const MESSAGE_KEYS = [
   'helpers.description',
   'helpers.categoryStyling',
   'helpers.categoryReact',
-  'helpers.categoryNumber',
   'helpers.common.importTitle',
   'helpers.common.usageTitle',
   'helpers.common.basicUsageTitle',
@@ -263,7 +262,6 @@ export const MESSAGE_KEYS = [
   'helpers.mergeProps.description',
   'helpers.chain.description',
   'helpers.createSafeContext.description',
-  'helpers.clamp.description',
   'theming.introduction',
   'theming.colorPaletteTitle',
   'theming.colorPaletteDescription',
