+++ packages/arte-odyssey/src/helpers/index.ts
@@ -1,5 +1,4 @@
 export { chain } from './chain';
-export { clamp } from './clamp';
 export { cn } from './cn';
 export { createSafeContext } from './create-safe-context';
 export { mergeProps } from './merge-props';
